+++ db4e/Constants/DElem.py
@@ -17,6 +17,7 @@ class DElem(ConstGroup):
     MONEROD : str = "monerod"
     MONEROD_REMOTE : str = "monerod_remote"
     P2POOL : str = "p2pool"
+    P2POOL_WATCHER : str = "p2pool_watcher"
     P2POOL_REMOTE : str = "p2pool_remote"
     XMRIG : str = "xmrig"
     XMRIG_REMOTE : str = "xmrig_remote"