+++ db4e/Constants/DMethod.py
@@ -21,6 +21,7 @@ class DMethod(ConstGroup):
     GET_NEW : str = "get_new"
     GET_DEPL : str = "get_deployment"
     GET_TUI_LOG : str = "get_tui_log"
+    GET_RUNTIME_LOG : str = "get_runtime_log"
     INITIAL_SETUP : str = "initial_setup"
     INITIAL_SETUP_PROCEED : str = "initial_setup_proceed"
     LOG_VIEWER : str = DField.LOG_VIEWER